#129 Continue refactoring/separating components
More test passing. Still WIP.
