+++ packages/ui-common/components/MultiAgentAccelerator/MultiAgentAccelerator.tsx
@@ -809,7 +809,7 @@ export const MultiAgentAccelerator: FC<MultiAgentAcceleratorProps> = ({
                             currentUser={username}
                             isAwaitingLlm={isAwaitingLlm}
                             isStreaming={isStreaming}
-                            isSelectedNetworkTemporary={isSelectedNetworkTemporary}
+                            isTemporaryNetwork={isSelectedNetworkTemporary}
                             networkDisplayName={networkDisplayName || ""}
                             networkId={isSelectedNetworkTemporary ? selectedNetwork : ""}
                             neuroSanURL={neuroSanURL}
